Stopwatch supports floats instead of long datatypes. Created Main class for MergeSort_vs_Quicksort

+++ stop_watch/Stopwatch.java
@@ -3,10 +3,10 @@ package stop_watch;
 public class Stopwatch {
 
     private boolean isCounting = false;
-    private long start;
-    private long stop;
+    private float start;
+    private float stop;
 
-    public long start() {
+    public float start() {
         if (!isCounting) {
             start = System.currentTimeMillis();
             isCounting = true;
@@ -15,7 +15,7 @@ public class Stopwatch {
         return -1;
     }
 
-    public long stop() {
+    public float stop() {
         if (isCounting) {
             stop = System.currentTimeMillis();
             isCounting = false;
@@ -24,9 +24,9 @@ public class Stopwatch {
         return -1;
     }
 
-    public long getTime() {
+    public float getTime() {
         if(isCounting){
-            long current = System.currentTimeMillis();
+            float current = System.currentTimeMillis();
             return current - start;
         }else
             return stop - start;
